requirements.txt
flask==2.3.2
spleeter==2.3.1
openai-whisper==20230419     
ffmpeg-python==0.2.0
google-genai==0.5.0          
werkzeug==2.3.7
